3 implement tmdbproxy

+++ src/main/java/com/example/moviematchapi/repository/GenreRepository.java
@@ -3,5 +3,5 @@ package com.example.moviematchapi.repository;
 import com.example.moviematchapi.model.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenreRepository extends JpaRepository<Genre, Long> {
+public interface GenreRepository extends JpaRepository<Genre, Integer> {
 }
